requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django>=2.2.4
django-background-tasks==1.2.0
django-ratelimit==2.0.0
ecdsa==0.13.2
future==0.17.1
gunicorn==19.9.0
idna==2.8
psycopg2-binary==2.8.2
pyasn1==0.4.5
python-jose==3.0.1
python-keycloak==0.16.0
python-keycloak-client==0.2.2
pytz==2019.1
requests==2.21.0
rsa==4.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
urllib3>=1.21.1,<1.25
factory-boy==2.12.0
Faker==2.0.1
